refactor multi_object_tracker
